added editSparePage class added ability to click on element from dropdown menu
