+++ akinfiievaProject/src/main/java/pages/SparePage.java
@@ -10,11 +10,6 @@ import parentPage.ParentPage;
 public class SparePage extends ParentPage {
     @FindBy(xpath = ".//div[@class='box-tools']//a[@class='btn btn-info btn-sm']")
     private WebElement buttonAdd;
-    @FindBy(xpath = ".//input[@id='spares_spareName']")
-    private WebElement inputSpareName;
-    @FindBy(xpath = ".//select[@id='spares_spareType']")
-    private WebElement dropDownList;
-
 
     public SparePage(WebDriver webDriver) {
         super(webDriver);
@@ -24,16 +19,4 @@ public class SparePage extends ParentPage {
         actionsWithOurElements.clickOnElement(buttonAdd);
     }
 
-    public void enterSpareName(String spareName) {
-        actionsWithOurElements.enterTextIntoInput(inputSpareName, spareName);
-    }
-
-//    public void selectSpareTypeFromDropDown(String spareTypeName) {
-//        actionsWithOurElements.selectItemFromDropDownList(dropDownList, spareTypeName);
-//    }
-
-    public void clickSpareTypeFromDropDownList(String spareTypeFromList) {
-        actionsWithOurElements.clickOnElement(dropDownList);
-        webDriver.findElement(By.xpath("//*[contains(text(),'"+spareTypeFromList+"')]")).click();
-    }
 }
