+++ akinfiievaProject/src/test/java/abstractParentTest/AbstractParentTest.java
@@ -5,6 +5,7 @@ import org.junit.Assert;
 import org.junit.Before;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
+import pages.EditSparePage;
 import pages.HomePage;
 import pages.LoginPage;
 import pages.SparePage;
@@ -17,6 +18,7 @@ public class AbstractParentTest {
     protected LoginPage loginPage;
     protected HomePage homePage;
     protected SparePage sparePage;
+    protected EditSparePage editSparePage;
 
     @Before
     public void setUp() {
@@ -28,6 +30,7 @@ public class AbstractParentTest {
         loginPage = new LoginPage(webDriver);
         homePage = new HomePage(webDriver);
         sparePage = new SparePage(webDriver);
+        editSparePage = new EditSparePage(webDriver);
     }
 
     @After
