+++ akinfiievaProject/src/test/java/loginTest/LoginWithPageObjectTest.java
@@ -2,7 +2,6 @@ package loginTest;
 
 import abstractParentTest.AbstractParentTest;
 import org.junit.Test;
-import pages.LoginPage;
 
 public class LoginWithPageObjectTest extends AbstractParentTest {
