+++ akinfiievaProject/src/test/java/sparesTests/SparesTests.java
@@ -3,21 +3,20 @@ package sparesTests;
 import abstractParentTest.AbstractParentTest;
 import org.junit.Test;
 
-import java.util.concurrent.TimeUnit;
 
 public class SparesTests extends AbstractParentTest {
     @Test
     public void addNewSpare() {
-        loginPage.fillingLoginFormAndSubmitIt("Student","909090");
+        loginPage.fillingLoginFormAndSubmitIt("Student", "909090");
         homePage.checkIsAvatarPresent();
         homePage.clickOnMenuDictionary();
         homePage.clickOnSubmenuSpares();
 
         sparePage.clickOnButtonAdd();
         final String spareName = "AkinfiievaSpare";
-        sparePage.enterSpareName(spareName);
-//        sparePage.selectSpareTypeFromDropDown("Датчики");
-        sparePage.clickSpareTypeFromDropDownList("Датчики");
+        editSparePage.enterSpareName(spareName);
+        //       editSparePage.selectSpareTypeFromDropDown("Датчики");
+        editSparePage.clickSpareTypeFromDropDownList("Датчики");
 
     }
 }
